fixing rounding+tests

+++ src/main/java/de/idrinth/gods_and_heroes/interfaces/God.java
@@ -1,11 +1,12 @@
 package de.idrinth.gods_and_heroes.interfaces;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface God extends Individual {
-    public int getBelieve();
+    public BigDecimal getBelieve();
     public List<Believer> getBelievers();
-    public int getRenown();
+    public BigDecimal getRenown();
     public boolean createWonder(Wonder wonder);
     public List<Hero> getHeroes();
     public List<Priest> getPriests();
