+++ src/main/java/de/idrinth/gods_and_heroes/interfaces/Mortal.java
@@ -1,8 +1,10 @@
 package de.idrinth.gods_and_heroes.interfaces;
 
+import java.math.BigDecimal;
+
 public interface Mortal extends Person {
-    public int getOnDeathRenown();
-    public int getOnDeathSouls();
-    public int getOnDeathBelieve();
+    public BigDecimal getOnDeathRenown();
+    public BigDecimal getOnDeathSouls();
+    public BigDecimal getOnDeathBelieve();
     public boolean isDead();
 }
